requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
pydantic==2.7.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
python-json-logger
dependency-injector
